Refactor column filter persistence methods and adjust imports
