+++ src/components/results/common/column-filter/persist-computation-column-filters.ts
@@ -1,25 +1,22 @@
 /**
- * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+import { UUID } from 'node:crypto';
 import { FilterConfig, TableType } from '../../../../types/custom-aggrid-types';
 import { updateComputationResultFiltersColumn } from '../../../../services/study/study-config';
-import { UUID } from 'node:crypto';
 
 export const persistComputationColumnFilters = (
-    filters?: FilterConfig[],
-    colId?: string,
-    studyUuid?: UUID,
-    tableType?: TableType,
-    filterSubType?: string,
-    onError?: (error: unknown) => void
+    studyUuid: UUID,
+    tableType: TableType,
+    tableId: string,
+    colId: string,
+    filters: FilterConfig[],
+    onError: (error: unknown) => void
 ) => {
-    if (!studyUuid || !colId || !filterSubType || !tableType) {
-        return;
-    }
-    const filter = filters?.find((f) => f.column === colId);
+    const filter = filters.find((f) => f.column === colId);
     const columnFilterInfos = {
         columnId: colId,
         columnFilterInfos: filter
@@ -31,7 +28,7 @@ export const persistComputationColumnFilters = (
               }
             : null,
     };
-    updateComputationResultFiltersColumn(studyUuid, tableType, filterSubType, columnFilterInfos).catch((error) => {
-        onError?.(error);
+    updateComputationResultFiltersColumn(studyUuid, tableType, tableId, columnFilterInfos).catch((error) => {
+        onError(error);
     });
 };
